+++ src/js/videojs.record.js
@@ -263,6 +263,24 @@
         },
 
         /**
+         * Show save as dialog in browser so the user can store the recorded media
+         * locally.
+         *
+         * @private
+         * @param {object} name - Object with names for the particular blob(s)
+         *     you want to save. File extensions are added automatically. For
+         *     example: {'video': 'name-of-video-file'}. Supported keys are
+         *     'audio', 'video' and 'gif'.
+         */
+        saveAs: function(name)
+        {
+            if (this.engine && name !== undefined)
+            {
+                this.engine.save(name);
+            }
+        },
+
+        /**
          * Invoked when recording is stopped and resulting stream is available.
          *
          * @private
@@ -1349,6 +1367,24 @@
                     // assign stream to audio/video element source
                     this.mediaElement.src = url;
                     break;
+            }
+        },
+
+        /**
+         * Show save as dialog in browser so the user can store the recorded media
+         * locally.
+         *
+         * @private
+         * @param {object} name - Object with one or more names for the particular
+         *     blob(s) you want to save. File extensions are added automatically.
+         *     For example: {'video': 'name-of-video-file'}. Supported keys are
+         *     'audio', 'video' and 'gif'.
+         */
+        saveAs: function(name)
+        {
+            if (this.engine && name !== undefined)
+            {
+                this.engine.saveAs(name);
             }
         },